# server/dashboard/src/app/core/api/model/deposit-page.model.ts
/**
 * Nrgyrent Dashboard API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DepositRow } from './deposit-row.model';


export interface DepositPage { 
    content: Array<DepositRow>;
    page: number;
    size: number;
    totalElements: number;
    totalPages: number;
}