# server/dashboard/src/app/core/api/model/deposit-row.model.ts
/**
 * Nrgyrent Dashboard API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * One row of the Deposits report. Amounts are decimal TRX and USDT.
 */
export interface DepositRow { 
    id: number;
    date: string;
    /**
     * Deposit amount in decimal TRX.
     */
    trxAmount: number;
    /**
     * Deposit amount in decimal USDT.
     */
    usdtAmount: number;
    balanceType: string;
    groupLogin: string | null;
    name: string | null;
    accountActivated: boolean;
    depositWallet: string | null;
    sender: string | null;
    status: string;
    transaction: string | null;
}