# server/dashboard/src/app/core/api/model/general-statistics.model.ts
/**
 * Nrgyrent Dashboard API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * General dashboard statistics. All values are decimal TRX.
 */
export interface GeneralStatistics { 
    /**
     * Profit for the selected period in decimal TRX.
     */
    periodProfit: number;
    /**
     * Current Bybit balance in decimal TRX.
     */
    bybitBalance: number;
    /**
     * AML profit for the selected period in decimal TRX.
     */
    periodAmlProfit: number;
    /**
     * Pending referral payments in decimal TRX.
     */
    pendingReferralPayments: number;
    /**
     * Current itrx.io balance in decimal TRX.
     */
    itrxBalance: number;
    /**
     * Current catfee.io balance in decimal TRX.
     */
    catfeeBalance: number;
    /**
     * Current trxx.io balance in decimal TRX.
     */
    trxxBalance: number;
    /**
     * Current netts.io balance in decimal TRX.
     */
    nettsBalance: number;
    /**
     * Sweep wallets balance in decimal TRX.
     */
    sweepWalletsBalance: number;
    /**
     * Sum of user balances in decimal TRX.
     */
    userBalancesTotal: number;
    /**
     * Referral payments for the selected period in decimal TRX.
     */
    periodReferralPayments: number;
    /**
     * Referral program remainder for the selected period in decimal TRX.
     */
    periodReferralProgramRemainder: number;
    /**
     * Sum of deposits in decimal TRX.
     */
    depositsTotal: number;
    /**
     * Provider commission for the selected period in decimal TRX.
     */
    periodProviderCommission: number;
    /**
     * New field value in decimal TRX.
     */
    newField: number;
}